Move dump_strings to lib and add benches
