+++ src/bin/pktstrings.rs
@@ -8,6 +8,7 @@ use std::path::Path;
 use std::vec::Vec;
 
 use pktstrings::net;
+use pktstrings::strings;
 
 const HELP_NUMBER: &str = "Number of printable characters to display";
 const HELP_BLOCK_PRINT: &str = "Print string blocks without packet info on each line";
@@ -65,85 +66,6 @@ fn apply_filter<T: Activated>(cap: &mut Capture<T>, bpf: &Option<String>, cmd: &
         match cap.filter(bpf, true) {
             Ok(_) => (),
             Err(err) => cmd.error(ErrorKind::InvalidValue, err).exit(),
-        }
-    }
-}
-
-/// Dump any valid ASCII strings over `len` size to stdout.
-///
-/// If `resolver` is passed in, this function will use it to perform DNS
-/// lookups on addresses.
-///
-/// If `block_print` is set, the function will only print the packet headers
-/// for the first ASCII line found.
-///
-/// If `regex` is provided this is run against the entire packet before
-/// the string scanning is run. Packets not matching the regex are skipped.
-fn dump_strings<T: Activated>(
-    cap: &mut Capture<T>,
-    len: &usize,
-    resolver: &mut Option<Box<net::Resolver>>,
-    block_print: &bool,
-    regex: &Option<Regex>,
-) {
-    let mut pkt_count = 0;
-
-    while let Ok(pkt) = cap.next_packet() {
-        pkt_count += 1;
-
-        if let Some(regex) = regex {
-            if !regex.is_match(pkt.data) {
-                continue;
-            }
-        }
-
-        let mut printed = false;
-        let mut chars = 0;
-        let mut partial = String::new();
-        let mut pkt_str: Option<String> = None;
-        for byte in pkt.data {
-            let c = *byte as char;
-            // TODO: other encodings
-            if c.is_ascii() && !c.is_ascii_control() {
-                chars += 1;
-                if chars > *len {
-                    print!("{}", c);
-                } else {
-                    partial.push(c);
-                    if chars == *len {
-                        if pkt_str.is_none() {
-                            if let Some(ref mut r) = resolver {
-                                let mut pktsum = net::PacketSummary::from_packet(&pkt, Some(r));
-                                pkt_str = Some(pktsum.formatted());
-                            } else {
-                                let mut pktsum = net::PacketSummary::from_packet(&pkt, None);
-                                pkt_str = Some(pktsum.formatted());
-                            }
-                        }
-
-                        let idx = pkt_count.to_string().blue();
-                        if !printed || !*block_print {
-                            if let Some(ref mut pkt_str) = pkt_str {
-                                print!("[{idx}]{pkt_str}: ");
-                                printed = true;
-                                if *block_print {
-                                    println!();
-                                }
-                            }
-                        }
-                        print!("{partial}");
-                    }
-                }
-            } else {
-                if chars >= *len {
-                    println!();
-                }
-                chars = 0;
-                partial.clear();
-            }
-        }
-        if chars >= *len {
-            println!();
         }
     }
 }
@@ -214,7 +136,7 @@ fn main() -> Result<(), clap::Error> {
         match Capture::from_file(file) {
             Ok(mut cap) => {
                 apply_filter(&mut cap, &cli.bpf_expression, &mut cmd);
-                dump_strings(
+                strings::dump_strings(
                     &mut cap,
                     &cli.number,
                     &mut resolver,
@@ -238,7 +160,7 @@ fn main() -> Result<(), clap::Error> {
             match capture_dev.open() {
                 Ok(mut cap) => {
                     apply_filter(&mut cap, &cli.bpf_expression, &mut cmd);
-                    dump_strings(
+                    strings::dump_strings(
                         &mut cap,
                         &cli.number,
                         &mut resolver,
@@ -264,7 +186,7 @@ fn main() -> Result<(), clap::Error> {
             Ok(inactive_cap) => match inactive_cap.immediate_mode(true).open() {
                 Ok(mut cap) => {
                     apply_filter(&mut cap, &cli.bpf_expression, &mut cmd);
-                    dump_strings(
+                    strings::dump_strings(
                         &mut cap,
                         &cli.number,
                         &mut resolver,
