+++ src/lib/lib.rs
@@ -1,2 +1,3 @@
 pub mod net;
 pub mod proto;
+pub mod strings;
